Fix qmldir copying in debug and release builds on Windows
This is a backport of 6cc02ce6c85d3dbd49a55060bd21a8359884786f from 5.7.

In a parallel build we may end up copying the qmldir file at the same
time, which doesn't work on Windows due to file locking. Apply the same
guard for the copying condition as in commit
770a0c91f3fadcdb132d9eb96d085aafbe1bacd0.

Task-number: QTBUG-57153
Change-Id: Ibac759b16cebaf04f5d2f785211b62071aa656a8
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>