doc: update QPrinter docs for clangqdoc
The docs for QPrinter contained a use of Q_QDoc to
declare an enum representing paper sizes for different
printers, but this special declaration for qdoc was
no longer needed, so it was removed. The real declaration
is in QPagedPaintDevice, which is the public base class
of QPrinter, and it is documented there. A few other
uses of Q_QDOC were examined and either removed or
upgraded to Q_CLANG_QDOC. One use of Q_OS_WIN was
changed to also be included for Q_CLANG_QDOC.

Change-Id: If6d810c624aa8d659fd0e3e753ba666d4d42ef83
Reviewed-by: Martin Smith <martin.smith@qt.io>