Add screen product information
Add new methods to QPlatformScreen that platform plugins can reimplement
in order to provide more information such as vendor, model and serial
number.

Expose that information as QScreen properties.

A use-case for this feature is a Wayland compositor that maps screens
to Wayland outputs hence it needs to replicate the information.

This information can also be added to the diagnostic output of qtdiag.

[ChangeLog][QtGui][QScreen] Add manufacturer, model and serialNumber
properties.

Change-Id: Ia6945f41023340602ef9d618e0d833a0c1825ab3
Reviewed-by: Johan Helsing <johan.helsing@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>