Re-fix build error with ICC 17 on Windows
This is a repeat of ae880beb7d02141c5097ef61409fa66b2c910dd3, which had
fixed the problem for ICC 16. That commit was a repeat of
acf80b9a2b913e898ed4c4ed14d4ea79401484fe, which had fixed it for ICC 15.

As reported in ae880beb7, ICC doesn't like polymorphic exported classes
with inline constructors. That commit added the default constructor, but
we forgot the copy constructor. This constructor should have been
protected, so users are forced to use the virtual clone() function, but
we can't make it so in Qt 5 because MSVC encodes the protection and has
exported the inline function in debug builds.

qsvgiconengine.obj : error LNK2001: unresolved external symbol "const QIconEngine::`vftable'" (??_7QIconEngine@@6)

Change-Id: I427336c52fc342638c74fffd149033b990ea7ade
Reviewed-by: Lars Knoll <lars.knoll@qt.io>